+++ dependencies.yaml
@@ -488,6 +488,7 @@ dependencies:
           - *numpy
           - python-louvain
           - scikit-learn>=0.23.1
+          - pylibwholegraph=23.10.*
   test_python_pylibcugraph:
     common:
       - output_types: [conda, pyproject]
